requirements.txt
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.1.6
djangorestframework==3.15.2
filelock==3.18.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
platformdirs==4.3.7
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
waitress==3.0.2
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
